wound system based off comp registery
