+++ Content.Shared/Wounds/Components/WoundBodyComponent.cs
