+++ Content.Shared/Wounds/Components/WoundComponent.cs
