+++ Content.Shared/Wounds/Components/WoundRecieverComponent.cs
